% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbVersion.R
\name{dbVersion}
\alias{dbVersion}
\title{Gets SQL DB Version}
\usage{
dbVersion(conn)
}
\arguments{
\item{conn}{A DBI Connection Object}
}
\description{
Given a DBI connection, evaluate the information to extract the DB Version.
This may be necessary when executing RDBMS-specific SQL syntax.
}
\examples{
db_con <- dbConnect(...)
dbVersion(db_con)
}
