% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTablePkey.R
\name{.dbTablePkey}
\alias{.dbTablePkey}
\title{Gets Primary Key from a SQL Table}
\usage{
.dbTablePkey(conn, name)
}
\arguments{
\item{conn}{A DBI Connection Object}

\item{name}{A table name in the DB}
}
\description{
Gets Primary Key from a SQL Table
}
