% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTablePkey_postgres.R
\name{.dbTablePkey_postgres}
\alias{.dbTablePkey_postgres}
\title{Gets Primary Key from Postgres Table}
\usage{
.dbTablePkey_postgres(conn, name)
}
\arguments{
\item{conn}{A DBI Connection Object}

\item{name}{A table name in the DB}
}
\description{
Gets Primary Key from Postgres Table
}
