% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUpsertStatement.R
\name{.dbUpsertStatement}
\alias{.dbUpsertStatement}
\title{Generate Upsert SQL Statement}
\usage{
.dbUpsertStatement(
  conn,
  target_table,
  staging_table,
  table_pkey,
  insert_cols,
  update_cols
)
}
\arguments{
\item{conn}{A DBI Connection Object}

\item{target_table}{A table name in the DB to upsert to}

\item{staging_table}{A table name in the DB containing data to upsert}

\item{table_pkey}{A character vector of column names used as the primary key}

\item{insert_cols}{A character vector of column names to insert, should include the PKs}

\item{update_cols}{A character vector of column names to update, should exclude the PKs}
}
\description{
Generate Upsert SQL Statement
}
