% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbColumnInfoExtended.R
\name{dbColumnInfoExtended}
\alias{dbColumnInfoExtended}
\title{Extended details about table columns}
\usage{
dbColumnInfoExtended(conn, name, schema, catalog)
}
\arguments{
\item{conn}{A DBI Connection Object}

\item{name}{A table name in the DB}

\item{schema}{A table schema in the DB}

\item{catalog}{A table catalog in the DB}
}
\description{
Extended details about table columns
}
