% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUpdateStatement.R
\name{.dbUpdateStatement}
\alias{.dbUpdateStatement}
\title{Generates update SQL Statement}
\usage{
.dbUpdateStatement(
  conn,
  target_table,
  staging_table,
  join_cols = NA,
  update_cols = NA
)
}
\arguments{
\item{conn}{A DBI Connection Object}

\item{target_table}{A table name in the DB to upsert to}

\item{staging_table}{A table name in the DB containing data to upsert}

\item{join_cols}{A character vector of column names used as the primary key}

\item{update_cols}{A character vector of column names to update, should exclude the PKs}
}
\description{
This function should never be called directly.
It is an internal function to the package.
Generates ANSI compliant SQL statement to update existing records from another table.
}
