% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUpsertStatement.R
\name{.dbUpsertStatement}
\alias{.dbUpsertStatement}
\title{Generate Upsert SQL Statement}
\usage{
.dbUpsertStatement(
  conn,
  target_table,
  staging_table,
  table_pkey,
  insert_cols,
  update_cols
)
}
\arguments{
\item{conn}{A DBI Connection Object}

\item{target_table}{A table name in the DB to upsert to}

\item{staging_table}{A table name in the DB containing data to upsert}

\item{table_pkey}{A character vector of column names used as the primary key}

\item{insert_cols}{A character vector of column names to insert, should include the PKs}

\item{update_cols}{A character vector of column names to update, should exclude the PKs}
}
\description{
This function should never be called directly.
It is an internal function to the package.
Generates ANSI compliant SQL statements to update existing records, and insert new records.
This function returns a list of 2 SQL statements.
The first element of text in the list is the SQL UPDATE statement.
The second element of text in the list is the SQL INSERT statement.
It is intended that both statements are sent in the same transaction.
}
